+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 - [minor] The length computation `actual_length / 2` for the random key will be done in the `KeyFactory` instead of the extension.
 
-- [minor] Removed the configuration option `password.phpass_iteration_length`. Instead the parameter `ma27_api_key_authentication.password_hasher.phpass.iteration_length` should be modified manually.
+- [breaking] Removed the configuration option `password.phpass_iteration_length`. Instead the parameter `ma27_api_key_authentication.password_hasher.phpass.iteration_length` should be modified manually.
+
+- [breaking] Removed the configuration option `services.password_hasher`. Create a custom hasher via DI tags instead.
 
 ## 1.2.0